Separate most networking code into a different file.
