+++ Network/IBeat.cs
